P3: keep raw directory ctype inputs unsigned
C ctype functions accept only EOF or values representable as unsigned char.
Ten checks consumed plain-char bytes originating from raw directory names, so
high-bit filenames could pass a negative value on signed-char targets. Cast
each byte to uint8_t at the ctype boundary, including indirect flows found by
helper, dictionary, callback, and retained-path tracing.

The affected callers are product hardware classification, EDAC controller and
DIMM discovery, NUMA chart discovery and cached NUMA counting, PCI AER
discovery, local-sockets PID scanning, journal remote-source labeling, and
journal writer-ID fallback parsing. ASCII names, accepted suffixes, PID
all-digit validation, journal metadata fallbacks, paths, return values, charts,
caching, recursion, error handling, ownership, locking, and cleanup remain
unchanged. High-bit bytes are now valid-domain ctype inputs instead of negative
values or accidental 0xff/EOF aliases.

The conversions add no I/O, allocation, loop, or algorithmic work and compile
to the existing integer argument path. They make behavior portable across
signed- and unsigned-char targets.

Validated every grammar with real readdir() entries containing ASCII, 0x80, and
0xff bytes under default, signed-char, unsigned-char, ASan, and UBSan builds.
Compiled every affected translation-unit context under both signed- and
unsigned-char modes using configured CMake commands, and linked the focused
systemd-journal.plugin target.

(cherry picked from commit 80513e2d004103151c9b70228b19f9a12ae3e0ef)

+++ src/libnetdata/local-sockets/local-sockets.h
@@ -617,7 +617,7 @@ static inline bool local_sockets_is_path_a_pid(const char *s) {
     if(!s || !*s) return false;
 
     while(*s) {
-        if(!isdigit(*s++))
+        if(!isdigit((uint8_t)*s++))
             return false;
     }
